backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
bcrypt==4.2.0
pyjwt==2.9.0
minio==7.2.9
httpx==0.27.2
aiofiles==24.1.0
python-frontmatter==1.1.0
